First Pass at Interpreter Architecture
Summary:
After about a week of experimentation and research, I made a first attempt at establishing an architecture for the interpretation side of this project. .csvs files are first processed into a parse tree by lark. Then, the tree is used to initialize a CslVisitor class/subclass, which visits the tree during initialization. The CslVisitor also maintains a stack used to store contextual information during the visit. As each node is visited, a set of stack operations are performed based on a corresponding method of the ExpressionIndex class. The result should be a single Schema object sitting in the stack, which is popped and aggregated within CslVisitor.

Additions:
+ lark grammar typo and bugfixing
+ demo/experimentation files
> test.csvs
> test_parse.py
+ CslVisitor
> Defines visitation behavior for .csvs parse trees
+ CslRuleset(CslVisitor)
> Will be used to contain .csvs opening and validation behavior
+ ExpressionIndex
> Singleton class which sreves as a container for methods which define stack manipulations during CslVisitor visits
>> As such, may later be wrapped into CslVisitor, but will wait to see how interpreter section develops
+ expresion_classes.py
> Defines classes which describe specific instances of nodes encountered during visit, as well as validation behavior

TODO
* Write out interpretation behavior for the rest of the CSL grammar
* Validation behavior
* Test suite
* GUI?