Additional Interpretation Behavior, Parenthesized
Summary:
Continued defining interpretation behavior for the CSL 1.2 grammar. Ran into a tricky issue with parenthesized expressions. Because these expressions are composed of an arbitrary number of column validation expressions, if a parenthesized expression is used to compose another expression which expects one or more column validation expressions in addition to the parenthesized expression, the parenthesized expression (if passed after another expression in sequence) has the potential to "eat" all the other validation expressions on the stack if we are simply popping the stack while its top is a validation expression. To work around this, when the visitor reaches a parenthesized expression, it recognizes this and only pops a number of elements of the stack equal to the number of the parenthesized expression's children.

Additions:
+ More interpretation behavior
+ ParenthesizedExpr workaround

TODO
* Write out interpretation behavior for the rest of the CSL grammar
* Validation behavior
* Test suite
* GUI?