Interpretation for External Expressions, Counted Expressions, Data-providing Expressions
Summary:
Interpretation behavior finished for data providing expressions and single external expressions. Data providing expression objects are now subclasses of the abstract DataExpr class, which demands that an .evaluate method be defined (for value production at runtime). In the process of finishing interpretation behavior for data-providing expressions, I realized that two of them could potentially cause the same issue as parenthesized expressions, in certain edge cases. As such, I generalized the handling strategy I created for parenthesized expressions to account for "counted expressions" as a whole - expressions which can only be properly interpreted using a stack-based model if the number of direct children is known at interpretation time.

Additions:
+ More interpretation behavior
+ Generalization of counted expressions solution
> CslVisitor now checks against a class variable counted_expressions to see if a given node must be counted. If so, calls _call_counted_expression instead of _call_expression during DFS
+ DataExpr
> An abstract class which is used to interface with expressions which dynamically provide data at runtime

TODO
* Interpretation behavior for
> Combinatorial Expressions
> Date expressions
> Conditional expressions
> Column directives
> Global directives
* Validation behavior
* Test suite
* GUI?