Switch to Explicit Context Definition using Recursive Stacks
Summary:
Use of a single shared stack for the interpreter was clever, but ultimately required too much wrenching to accommodate expressions which defined their own sub-contexts (such as parentheticals and some string provider children which take string providers as their own children). To avoid a hardcoded check for such expressions, which would be unacceptably fragile, I've switched to a model in which sub-stacks are recursively generated for each expression by calling .visit on each child.

Additions:
+ Recursive stack generation
> Each ._call_expression target takes a stack as an argument
> The stack is generated within .visit by recursively calling .visit on each child of the current node
> This delinaeates each expression's context while maintaining a single unified visit behavior, instead of defining that per-expression
+ Adjusted lark representation of CSL 1.2 grammar to allow for the extraction of certain values, like version numbers

TODO
* Interpretation behavior for
> Combinatorial Expressions
> Date expressions
> Conditional expressions
> Column directives
> Global directives
* Validation behavior
* Test suite
* GUI?